requirements.txt
cycler==0.10.0
kiwisolver==1.1.0
matplotlib
numpy==1.17.2
pandas==0.25.1
Pillow==8.1.1
pykitti==0.3.1
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.3
scipy==1.3.1
six==1.12.0
torch==1.4.0
torchsummary==1.5.1
torchvision==0.5.0
